% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BH_q.R
\name{BH_q}
\alias{BH_q}
\title{Generate q-values based on the BH (Benjamini and Hochberg) approach}
\usage{
BH_q(pvals)
}
\arguments{
\item{pvals}{vector of observed p-values.}
}
\description{
Benjamini and Hochberg (1995) developed the first approach for controlling FDR. Unlike the approach proposed by Millstein and Volfson (2013), it is based on the idea that a significance level (alpha) is specified a priori by the investigator. Storey and Tibshirani (2003) developed the q-value concept, where FDR is estimated at each observed p-value.
}
\examples{
nrow_=100
ncol_=100
X = as.data.frame(matrix(rnorm(nrow_*ncol_),nrow=nrow_,ncol=ncol_))
Y = as.data.frame(matrix(rnorm(nrow_*ncol_),nrow=nrow_,ncol=ncol_))
nperm = 10

myanalysis = function(X,Y){
	ntests = ncol(X)
	rslts = as.data.frame(matrix(NA,nrow=ntests,ncol=2))
	names(rslts) = c("ID","pvalue")
	rslts[,"ID"] = 1:ntests
	for(i in 1:ntests){
		fit = cor.test(X[,i],Y[,i],na.action="na.exclude",
			alternative="two.sided",method="pearson")
		rslts[i,"pvalue"] = fit$p.value
	}
	return(rslts)
} # End myanalysis

# Generate observed results
obs = myanalysis(X,Y)
q.values.BH = BH_q( obs[, "pvalue"] )

}
\references{
Benjamini, Yoav, and Yosef Hochberg. "Controlling the false discovery rate: a practical and powerful approach to multiple testing." Journal of the royal statistical society. Series B (Methodological) (1995): 289-300.

Storey, John D., and Robert Tibshirani. "Statistical significance for genomewide studies." Proceedings of the National Academy of Sciences 100.16 (2003): 9440-9445.
}
\author{
Joshua Millstein
}
