% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meff.R
\name{meff.jm}
\alias{meff.jm}
\title{Estimate the Effective Number of Tests}
\usage{
meff.jm(mydat, B = 1, seed)
}
\arguments{
\item{mydat}{A design matrix with \eqn{n} observations (rows) and \eqn{p} covariates (columns).}

\item{B}{Integer. Number of permutations to perform. (Default is 1)}

\item{seed}{Integer. Setting the seed for reproducibility.}
}
\value{
Numeric. Returns the estimated effective number of tests averaged over \code{B} permutations.
}
\description{
Estimate the effective number of tests using a permutation-based approach.
}
\examples{

# Independent
ss=100
nvar=100
X = as.data.frame(matrix(rnorm(ss * nvar), nrow = ss, ncol = nvar))
meff.jm(X, B = 5, seed = 1234)

# High correlation
S = matrix(0.9, ss, nvar)
diag(S) = 1
X = as.matrix(X) \%*\% chol(S)
meff.jm(X, B = 5, seed = 1234)
}
\references{
Millstein J, Volfson D. 2013. Computationally efficient
permutation-based confidence interval estimation for tail-area FDR.
Frontiers in Genetics | Statistical Genetics and Methodology 4(179):1-11.
}
\author{
Joshua Millstein, Eric S. Kawaguchi
}
